% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-bingham-3d.R
\name{bingham.test3d}
\alias{bingham.test3d}
\title{Test of isotropy of antipodally symmetric data}
\usage{
bingham.test3d(x)
}
\arguments{
\item{x}{unit vectors, [x1,...,xn]}
}
\description{
Test of isotropy of antipodally symmetric data
}

