% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confint.ellipsoid}
\alias{confint.ellipsoid}
\title{Confidence interval for ellipsoid parameters using simulation}
\usage{
\method{confint}{ellipsoid}(x, fun, nsim = 1000, probs = c(0.025, 0.975),
  ...)
}
\arguments{
\item{tol}{tolerance for absolute  deviation in the ||beta||=1 constraint.}
}
\description{
Assuming normality of errors.
}

