% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipse_contrasts.R
\name{ellipse_contrast_2d}
\alias{ellipse_contrast_2d}
\title{Testing contrast for ellipses}
\usage{
ellipse_contrast_2d(els, out = TRUE)
}
\description{
Give this to confint.ellipsoid to compute the contrast for
 testing semi_a - semi_b = 0
}

