% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ellipse_form}
\alias{ellipse_form}
\title{Ellipse center and matrix from general parameter form}
\usage{
ellipse_form(beta, d)
}
\arguments{
\item{beta}{OLS estimates}

\item{d}{dimension}
}
\description{
Ellipse center and matrix from general parameter form
}

