% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ellipse_solve_rota}
\alias{ellipse_solve_rota}
\title{Solve rotation and semi-axes from general transform}
\usage{
ellipse_solve_rota(A)
}
\arguments{
\item{A}{the trasformation matrix in ellipsoid equation}
}
\description{
Solve rotation and semi-axes from general transform
}

