% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid_from_beta}
\alias{ellipsoid_from_beta}
\title{convert beta vector to ellipsoid object}
\usage{
ellipsoid_from_beta(beta, d)
}
\arguments{
\item{beta}{beta vector, the coefficients in quadratic form}

\item{d}{dimension}
}
\description{
convert beta vector to ellipsoid object
}

