% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid_OLS}
\alias{ellipsoid_OLS}
\title{OLS fitting of an ellipsoid to d-dimensional point cloud}
\usage{
ellipsoid_OLS(x, origin = FALSE)
}
\arguments{
\item{x}{point coordinate matrix}

\item{origin}{Fix center of the ellipsoid to origin?}
}
\description{
OLS fitting of an ellipsoid to d-dimensional point cloud
}

