% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid_shape}
\alias{ellipsoid_shape}
\title{Ellipsoid shape for 3d plotting}
\usage{
ellipsoid_shape(N = 2, axes = c(1, 1, 1), R = NULL)
}
\arguments{
\item{R}{Rotation matrix.}
}
\description{
Refine an icosahedron, then transform
}

