% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Euler2quaternion}
\alias{Euler2quaternion}
\title{Euler axis/angle to quaternion}
\usage{
Euler2quaternion(euler)
}
\arguments{
\item{euler}{c(e1,e2,e3,angle) where e* is the unit axis of rotation}
}
\description{
Euler axis/angle to quaternion
}

