% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_intensity.R
\name{intensity.sphere}
\alias{intensity.sphere}
\title{intensity on a sphere}
\usage{
intensity.sphere(latlon, s = 0.25, nlat = 20, newlatlon, plot = FALSE,
  points = FALSE, ...)
}
\arguments{
\item{latlon}{observations}

\item{s}{sd for gaussian kernel}

\item{nlat}{if we make a grid, use nlat x (2*lat)}

\item{newlatlon}{if given predict here, else on grid}

\item{plot}{should we plot (only with grid)}

\item{points}{add the points to the plot}
}
\description{
Gaussian kernel
}

