% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mean_ellipsoids}
\alias{mean_ellipsoids}
\title{Mean ellipsoid using empirical mean}
\usage{
mean_ellipsoids(x, nsim = 1000, weight, just_rotation = FALSE,
  add_noise = FALSE, keep_data = FALSE, ...)
}
\arguments{
\item{x}{a list of ellipsoids}

\item{nsim}{Total number of points to simulate from ellipsoids for average estimation}

\item{weight}{Weigth the ellipsoid samples. Default is to use standard deviation of the OLS estimates.}

\item{just_rotation}{Rescale each ellipse to have same determinant?}

\item{add_noise}{Simulate noise using the s2 ols estimate?}
}
\description{
The sample size is proportional to estimated s2
}

