% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipsoid-mean.R
\name{mean_rotation_quat}
\alias{mean_rotation_quat}
\title{Mean ellipsoid rotation using quaternion average}
\usage{
mean_rotation_quat(x, ...)
}
\description{
Mean ellipsoid rotation using quaternion average
}

