% Generated by roxygen2 (4.0.2): do not edit by hand
\name{persp.ellipsoid}
\alias{persp.ellipsoid}
\title{Perspective plot of an ellipsoid in 3d}
\usage{
\method{persp}{ellipsoid}(x, add = FALSE, theta = 25, phi = 30,
  expand = 0.9, triptych = FALSE, pmat, ...)
}
\description{
Perspective plot of an ellipsoid in 3d
}

