% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{plot.ellipsoid}
\alias{plot.ellipsoid}
\title{Plot an ellipsoid}
\usage{
\method{plot}{ellipsoid}(x, add = TRUE, persp3d = FALSE, res = 201,
  scale = 1, ...)
}
\description{
Plot an ellipsoid
}

