% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sphere.smooth.R
\name{plot.spheresmooth}
\alias{plot.spheresmooth}
\title{plot the 3d sphere of predictions}
\usage{
\method{plot}{spheresmooth}(obj, col = heat.colors, col_zlim, ...,
  data = FALSE)
}
\description{
plot the 3d sphere of predictions
}

