% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quaternion2EulerAngles}
\alias{quaternion2EulerAngles}
\title{Quaternion to Euler angles}
\usage{
quaternion2EulerAngles(q)
}
\description{
from euclideanspace.com,
}
\details{
Convention: Euler is (Heading, Attitude, Bank) so that
Heading = rotation  around y, Attitude = rotation around z,
Bank = rotation around x, in that order.
}

