% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rellipsoid}
\alias{rellipsoid}
\title{Uniform sample on the surface of an ellipsoid}
\usage{
rellipsoid(n, axes = c(1, 1, 1), noise = 0, R = NULL)
}
\arguments{
\item{R}{rotation matrix}
}
\description{
Uniform sample on the surface of an ellipsoid
}

