% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{rotationMatrix2Euler}
\alias{rotationMatrix2Euler}
\title{Rotation matrix (prop or improp) to Euler axis-angle}
\usage{
rotationMatrix2Euler(R)
}
\description{
Rotation matrix (prop or improp) to Euler axis-angle
}

