% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coords.R
\name{rotationMatrix2EulerAngles}
\alias{rotationMatrix2EulerAngles}
\title{Rotation matrix to Euler angles}
\usage{
rotationMatrix2EulerAngles(R)
}
\value{
(Heading, Attitude, Bank)
}
\description{
Rotation matrix to Euler angles
}

