% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runifcircle}
\alias{runifcircle}
\title{Sample unit circle uniformly}
\usage{
runifcircle(n, polar = FALSE)
}
\arguments{
\item{n}{sample size}

\item{polar}{Return polar coordinates? (Def: FALSE)}
}
\description{
Sample unit circle uniformly
}
\details{
Default is to return unit vectors.
}

