% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runifsphere}
\alias{runifsphere}
\title{Sample unit sphere uniformly}
\usage{
runifsphere(n, spherical = FALSE)
}
\arguments{
\item{n}{sample size}

\item{spherical}{Return azimuth-inclination? (Def: FALSE)}
}
\description{
Sample unit sphere uniformly
}
\details{
Default is to return unit vectors.
}

