% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{sample_ellipse_beta}
\alias{sample_ellipse_beta}
\title{Sample from the OLS estimate of the beta parameters}
\usage{
sample_ellipse_beta(x, nsim = 100, tol = 0, maxiter = 5000)
}
\description{
Sample from the OLS estimate of the beta parameters
}

