% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.grid.R
\name{sphere.grid}
\alias{sphere.grid}
\title{generate grids on sphere}
\usage{
sphere.grid(N = 2, lower = TRUE, ico = TRUE)
}
\arguments{
\item{N}{number of refinenemts}

\item{lower}{return also the lower hemishpere?}

\item{ico}{use triangulation based on icosahedron (req. rgl)? N is then takes as refinement steps, should be <5!

#' Two ways:

  A) a decreasing radius ring of N points from equator to pole with alternating 2pi/N shift. 
  Concentrates on the poles, useful for statitical estimation if the pole area is where things happen.
  
  B) Triangulation by refining an icosahedron.
  
Returns latitude-longitude matrix. use ll2xyz or ll2ai for transformation.}
}
\description{
generate grids on sphere
}
