% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-uniform-latlon.R
\name{test.unifsphere}
\alias{test.unifsphere}
\title{Test of uniformity on a sphere}
\usage{
test.unifsphere(ai)
}
\arguments{
\item{ai}{azimuth-inclination matrix (n x 2)}
}
\description{
Independent test of uniformity in azimuth and inclination.
}
\details{
Returns the p value of Chi2 test for both angles
}

