% Generated by roxygen2 (4.0.2): do not edit by hand
\name{values2colors}
\alias{values2colors}
\title{values to colors}
\usage{
values2colors(v, n = 100, zlim, col = heat.colors, ...)
}
\arguments{
\item{v}{the values}

\item{n}{number of colors}

\item{zlim}{limits}

\item{color}{function, e.g. heat.colors, gray.colors}
}
\description{
values to colors
}

