% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coords.R
\name{xyz2globe}
\alias{xyz2globe}
\title{OLD for plotting on the globe, we need to trasform as the texture is fuckd up}
\usage{
xyz2globe(xyz)
}
\description{
OLD for plotting on the globe, we need to trasform as the texture is fuckd up
}

