% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ellipsoid_from_beta}
\alias{ellipsoid_from_beta}
\title{convert beta vector to ellipsoid object}
\usage{
ellipsoid_from_beta(beta, d)
}
\arguments{
\item{beta}{beta vector, the coefficients in quadratic form}

\item{d}{dimension}
}
\description{
convert beta vector to ellipsoid object
}

