% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.ellipsoid}
\alias{predict.ellipsoid}
\title{Predict i.e. give the length of a direction to be on the ellipsoid}
\usage{
\method{predict}{ellipsoid}(x, u, ...)
}
\description{
Returns the distance from ellipsoid center to the ellipsoid surface in
the given directions.
}

