% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sphere.predict}
\alias{sphere.predict}
\title{Predict on sphere using gaussian regression (Kriging)}
\usage{
sphere.predict(latlon, v, newlatlon, s = 0.25)
}
\arguments{
\item{latlon}{latitude-longitude matrix}

\item{v}{value at locations. if not given, we do intensity estimation.}

\item{newlatlon}{new locations}

\item{s}{gaussian kernel sd

use xyz2ll or ai2ll to transform from unitvecs or (azimuth, inclination) (physical)}
}
\description{
if no v is given, compute intensity.
}

