% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sphere.smooth.R
\name{sphere.smooth}
\alias{sphere.smooth}
\title{Smoothed unit sphere values}
\usage{
sphere.smooth(latlon, v, N = 3, s = 0.25)
}
\arguments{
\item{latlon}{latitude-longitude of data}

\item{v}{values on data points}

\item{N}{refinenement of the icosahedron triangulation}

\item{s}{smoothing sd for gaussian smoothing}
}
\description{
Krige on a sphere, return a plottable orb.
}

